+++ Projects/Server.Tests/Network/Packets/Old/Outgoing/DamagePacketTests.cs
@@ -18,7 +18,8 @@ namespace Server.Tests.Network.Packets
     {
       Span<byte> data = new DamagePacketOld(mobile, inputAmount).Compile();
 
-      Span<byte> expectedData = stackalloc byte[] {
+      Span<byte> expectedData = stackalloc byte[]
+      {
         0xBF, // Packet
         0x00, 0x0B, // Length
         0x00, 0x22, // Sub-packet
@@ -40,7 +41,8 @@ namespace Server.Tests.Network.Packets
     {
       Span<byte> data = new DamagePacket(mobile, inputAmount).Compile();
 
-      Span<byte> expectedData = stackalloc byte[] {
+      Span<byte> expectedData = stackalloc byte[]
+      {
         0x0B, // Packet
         0x00, 0x00, 0x00, 0x00, // Mobile Serial
         0x00, 0x00 // Amount
