+++ Projects/Server.Tests/Network/Packets/Old/Outgoing/TestEquipmentPackets.cs
