+++ Projects/Server/Mobile.cs
@@ -889,7 +889,7 @@ namespace Server
           ++m_ChangingCombatant;
           m_Combatant = value;
 
-          if ((m_Combatant != null && !CanBeHarmful(m_Combatant, false)) ||
+          if (m_Combatant != null && !CanBeHarmful(m_Combatant, false) ||
               !Region.OnCombatantChange(this, old, m_Combatant))
           {
             m_Combatant = old;
