+++ Projects/Server/Network/Packets.cs
@@ -29,7 +29,6 @@ using Server.Accounting;
 using Server.Compression;
 using Server.ContextMenus;
 using Server.Gumps;
-using Server.HuePickers;
 using Server.Items;
 using Server.Menus;
 using Server.Menus.ItemLists;
@@ -56,133 +55,6 @@ namespace Server.Network
     TryToSteal = 3,
     AreHolding = 4,
     Inspecific = 5
-  }
-
-  public sealed class ChangeUpdateRange : Packet
-  {
-    private static readonly ChangeUpdateRange[] m_Cache = new ChangeUpdateRange[0x100];
-
-    public ChangeUpdateRange(int range) : base(0xC8, 2)
-    {
-      Stream.Write((byte)range);
-    }
-
-    public static ChangeUpdateRange Instantiate(int range)
-    {
-      var idx = (byte)range;
-      var p = m_Cache[idx];
-
-      if (p == null)
-      {
-        m_Cache[idx] = p = new ChangeUpdateRange(range);
-        p.SetStatic();
-      }
-
-      return p;
-    }
-  }
-
-  public sealed class ChangeCombatant : Packet
-  {
-    public ChangeCombatant(Mobile combatant) : base(0xAA, 5)
-    {
-      Stream.Write(combatant?.Serial ?? Serial.Zero);
-    }
-  }
-
-  public sealed class DisplayHuePicker : Packet
-  {
-    public DisplayHuePicker(HuePicker huePicker) : base(0x95, 9)
-    {
-      Stream.Write(huePicker.Serial);
-      Stream.Write((short)0);
-      Stream.Write((short)huePicker.ItemID);
-    }
-  }
-
-  public sealed class UnicodePrompt : Packet
-  {
-    public UnicodePrompt(Prompt prompt) : base(0xC2)
-    {
-      EnsureCapacity(21);
-
-      Stream.Write(prompt.Serial);
-      Stream.Write(prompt.Serial);
-      Stream.Write(0);
-      Stream.Write(0);
-      Stream.Write((short)0);
-    }
-  }
-
-  public sealed class ChangeCharacter : Packet
-  {
-    public ChangeCharacter(IAccount a) : base(0x81)
-    {
-      EnsureCapacity(305);
-
-      var count = 0;
-
-      for (var i = 0; i < a.Length; ++i)
-        if (a[i] != null)
-          ++count;
-
-      Stream.Write((byte)count);
-      Stream.Write((byte)0);
-
-      for (var i = 0; i < a.Length; ++i)
-        if (a[i] != null)
-        {
-          var name = a[i].Name;
-
-          if (name == null)
-            name = "-null-";
-          else if ((name = name.Trim()).Length == 0)
-            name = "-empty-";
-
-          Stream.WriteAsciiFixed(name, 30);
-          Stream.Fill(30); // password
-        }
-        else
-        {
-          Stream.Fill(60);
-        }
-    }
-  }
-
-  public sealed class DeathStatus : Packet
-  {
-    public static readonly Packet Dead = SetStatic(new DeathStatus(true));
-    public static readonly Packet Alive = SetStatic(new DeathStatus(false));
-
-    public DeathStatus(bool dead) : base(0x2C, 2)
-    {
-      Stream.Write((byte)(dead ? 0 : 2));
-    }
-
-    public static Packet Instantiate(bool dead) => dead ? Dead : Alive;
-  }
-
-  public sealed class SpeedControl : Packet
-  {
-    public static readonly Packet WalkSpeed = SetStatic(new SpeedControl(2));
-    public static readonly Packet MountSpeed = SetStatic(new SpeedControl(1));
-    public static readonly Packet Disable = SetStatic(new SpeedControl(0));
-
-    public SpeedControl(int speedControl)
-      : base(0xBF)
-    {
-      EnsureCapacity(3);
-
-      Stream.Write((short)0x26);
-      Stream.Write((byte)speedControl);
-    }
-  }
-
-  public sealed class InvalidMapEnable : Packet
-  {
-    public InvalidMapEnable() : base(0xC6, 1)
-    {
-    }
   }
 
   public sealed class BondedStatus : Packet
