+++ Projects/Server/Network/Packets/Old Packets/AccountPackets.cs
@@ -2,7 +2,7 @@
  * ModernUO                                                              *
  * Copyright (C) 2019-2020 - ModernUO Development Team                   *
  * Email: hi@modernuo.com                                                *
- * File: MobilePackets.cs - Created: 2020/05/07 - Updated: 2020/05/07    *
+ * File: AccountPackets.cs - Created: 2020/05/08 - Updated: 2020/05/08   *
  *                                                                       *
  * This program is free software: you can redistribute it and/or modify  *
  * it under the terms of the GNU General Public License as published by  *
@@ -18,15 +18,42 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
  *************************************************************************/
 
+using Server.Accounting;
+
 namespace Server.Network
 {
-  public sealed class DeathAnimation : Packet
+  public sealed class ChangeCharacter : Packet
   {
-    public DeathAnimation(Mobile killed, Item corpse) : base(0xAF, 13)
+    public ChangeCharacter(IAccount a) : base(0x81)
     {
-      Stream.Write(killed.Serial);
-      Stream.Write(corpse?.Serial ?? Serial.Zero);
-      Stream.Write(0);
+      EnsureCapacity(305);
+
+      var count = 0;
+
+      for (var i = 0; i < a.Length; ++i)
+        if (a[i] != null)
+          ++count;
+
+      Stream.Write((byte)count);
+      Stream.Write((byte)0);
+
+      for (var i = 0; i < a.Length; ++i)
+        if (a[i] != null)
+        {
+          var name = a[i].Name;
+
+          if (name == null)
+            name = "-null-";
+          else if ((name = name.Trim()).Length == 0)
+            name = "-empty-";
+
+          Stream.WriteAsciiFixed(name, 30);
+          Stream.Fill(30); // password
+        }
+        else
+        {
+          Stream.Fill(60);
+        }
     }
   }
 }
